+++ CHANGELOG.md
@@ -472,6 +472,45 @@ In this example:
 -   The method will first check `prop1`, then move to `prop2`, and finally `prop3`.
 -   If any of these keys do not exist, it will return `null`.    
 
+### **New Feature: `mergeWith()` Method**
+
+We’ve also added a powerful new method called **`mergeWith()`**, which allows you to merge two `MagicObject` instances with ease.
+
+#### What it Does:
+
+-   Combines properties from another object into the current one.
+-   If a property doesn’t exist in the current object, it will be added.
+-   If the property already exists:
+    -   If both values are `MagicObject` instances, they will be **merged recursively**.
+    -   Otherwise, the value will be **overwritten**.
+
+#### Example:
+
+```php
+$obj1 = new  MagicObject();
+$obj1->loadYamlString('
+user:
+  name: ALice
+client:
+  address: Jakarta
+', false, true, true);
+
+$obj2 = new  MagicObject();
+$obj2->loadYamlString('
+user:
+  email: alice@example.com
+client:
+  name: Ana
+', false, true, true);
+$obj1->mergeWith($obj2);
+
+echo  $obj1;
+
+// {"user":{"name":"ALice","email":"alice@example.com"},"client":{"address":"Jakarta","name":"Ana"}}
+```
+
+This method simplifies combining nested objects and ensures consistency in structured data merging.
+
 ### **Bug Fixes & Performance Improvements**
 
 -   Various small bug fixes related to edge cases.